+++ NEWS.md
@@ -80,6 +80,9 @@ sufficient information, see the ChangeLog file or Redmine
     def square(x) = x * x
     ```
 
+* Interpolated String literals are no longer frozen when
+  `# frozen-string-literal: true` is used. [[Feature #17104]]
+
 ## Command line options
 
 ### `--help` option
@@ -373,6 +376,7 @@ Excluding feature bug fixes.
 [Feature #16746]: https://bugs.ruby-lang.org/issues/16746
 [Feature #16754]: https://bugs.ruby-lang.org/issues/16754
 [Feature #16828]: https://bugs.ruby-lang.org/issues/16828
+[Feature #17104]: https://bugs.ruby-lang.org/issues/17104
 [Misc #16961]:    https://bugs.ruby-lang.org/issues/16961
 [Feature #17122]: https://bugs.ruby-lang.org/issues/17122
 [GH-2991]:        https://github.com/ruby/ruby/pull/2991